backend/requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
djongo==1.3.6
dnspython==2.2.1
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==2.4.0
pymongo==3.12.0
python-snappy==0.6.1
pytz==2022.1
sqlparse==0.2.4
tzdata==2022.1
winkerberos==0.9.0